chore: moving new navigation files
