+++ packages/application-shell/src/components/navbar/navbar-skeleton.tsx
