+++ packages/application-shell/src/components/navbar/navbar.mod.css
