+++ packages/application-shell/src/components/navbar/navbar.styles.d.ts
